simplify init/running,support  interactive usage
