+++ lightningbringer/__init__.py
@@ -1,1 +1,3 @@
 from .system import System
+from .config import init_config
+from .runner import init_system, init_trainer
